refactor: lab 패키지 이동
