+++ src/main/java/org/codequistify/master/application/lab/dto/PShellCreateRequest.java
@@ -1,0 +1,4 @@
+package org.codequistify.master.application.lab.dto;
+
+public record PShellCreateRequest() {
+}
