+++ src/main/java/org/codequistify/master/application/lab/dto/PShellCreateResponse.java
@@ -1,4 +1,4 @@
-package org.codequistify.master.core.domain.lab.dto;
+package org.codequistify.master.application.lab.dto;
 
 public record PShellCreateResponse(
         String url,
