--- src/main/java/org/codequistify/master/core/domain/lab/dto/PShellCreateRequest.java
@@ -1,4 +1,0 @@
-package org.codequistify.master.core.domain.lab.dto;
-
-public record PShellCreateRequest() {
-}
