+++ src/main/java/org/codequistify/master/core/domain/stage/service/impl/StageManagementServiceImpl.java
@@ -3,8 +3,8 @@ package org.codequistify.master.core.domain.stage.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.codequistify.master.application.exception.ApplicationException;
 import org.codequistify.master.application.exception.ErrorCode;
+import org.codequistify.master.application.lab.service.LabAssignmentService;
 import org.codequistify.master.application.player.service.PlayerProfileService;
-import org.codequistify.master.core.domain.lab.service.LabAssignmentService;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.core.domain.stage.convertoer.QuestionConverter;
 import org.codequistify.master.core.domain.stage.convertoer.StageConverter;
